bot/requirements.txt
discord.py
openai==1.51.2
psycopg2-binary
python-dotenv
redis
Flask-Login
requests==2.26.0